\name{KGE}
\alias{KGE}
\title{KGE (Kling-Gupta Efficiency)}
\usage{
KGE(model, observation, na.rm=TRUE, s.r=1, s.alpha=1, s.beta=1)
}
\description{
Calculate the Kling-Gupta Efficiency - between model and observation
}
\examples{


KGE(model, observation, na.rm=TRUE, s.r=1, s.alpha=1, s.beta=1)

}
\arguments{
\item{model}{model streamflow time series}

\item{observation}{observation streamflow time series }

}
\value{
Retun Kling-Gupta Efficiency between model and observation.
}

