\name{NSEwt}
\alias{NSEwt}
\title{Weighted NSE (Nash-Sutcliffe efficiency)}
\usage{
NSEwt(model, observation, w=0.5, p=1)
}
\description{
Weighted Nash-Sutcliffe efficiency- between model and observation
}
\examples{

NSEwt(model, observation, w=0.5, p=1)

}
\arguments{
\item{model}{model streamflow time series}

\item{observation}{observation streamflow time series }

}
\value{
Retun Weighted Nash-Sutcliffe efficiency between model and observation.
}


