\name{PBias}
\alias{PBias}
\title{PBias (Percentage Bias)}
\usage{
PBias(model,observation,na.rm=TRUE)
}
\description{
Percentage Bias- between model and observation
}
\examples{

PBias(mod,obs)

}
\arguments{
\item{model}{model streamflow time series}

\item{observation}{observation streamflow time series }

}
\value{
Retun percentage between model and observation.
}


