\name{RSquared}
\alias{RSquared}
\title{RSquared (Coefficient of Determination)}
\usage{
RSquared(model, observation)
}
\description{
Calculate the Coefficient of Determination - between model and observation
}
\examples{


RSquared(model, observation)

}
\arguments{
\item{model}{model streamflow time series}

\item{observation}{observation streamflow time series }

}
\value{
Retun Coefficient of Determination between model and observation.
}

